fix: resolve remaining CI warnings across GCC and clang-tidy
- art_internal_impl.hpp: cast (i+1) to unsigned in bitmask shift (sign-conversion)
- art.hpp: suppress GCC suggest-attribute=noreturn on insert_internal_fixed
- test_art_key_view_full_chain.cpp: uppercase U suffix, NOLINT constant seed,
  remove useless size_t→uint64_t casts, use .empty() not .size()>=1
- benchmark/micro_benchmark_utils.hpp: remove useless size_t→uint64_t casts

+++ benchmark/micro_benchmark_utils.hpp
@@ -314,10 +314,7 @@ class key_view_set {
     ks.views_.reserve(n);
     unodb::key_encoder enc;
     for (std::size_t i = 0; i < n; ++i) {
-      const auto kv = enc.reset()
-                          .encode(tag)
-                          .encode(static_cast<std::uint64_t>(i))
-                          .get_key_view();
+      const auto kv = enc.reset().encode(tag).encode(i).get_key_view();
       UNODB_DETAIL_DISABLE_MSVC_WARNING(26459)
       UNODB_DETAIL_DISABLE_MSVC_WARNING(26481)
       std::copy(kv.begin(), kv.end(), ks.buf_.data() + i * 9);
@@ -344,7 +341,7 @@ class key_view_set {
                           .encode(tag)
                           .encode(std::uint64_t{0x4242424242424242ULL})
                           .encode(static_cast<std::uint8_t>(i & 0xFF))
-                          .encode(static_cast<std::uint64_t>(i))
+                          .encode(i)
                           .get_key_view();
       UNODB_DETAIL_DISABLE_MSVC_WARNING(26481)
       std::ranges::copy(kv, ks.buf_.data() + i * 18);
@@ -428,8 +425,7 @@ class key_view_set {
     ks.views_.reserve(n);
     unodb::key_encoder enc;
     for (std::size_t i = 0; i < n; ++i) {
-      const auto kv =
-          enc.reset().encode(static_cast<std::uint64_t>(i)).get_key_view();
+      const auto kv = enc.reset().encode(i).get_key_view();
       UNODB_DETAIL_DISABLE_MSVC_WARNING(26481)
       std::ranges::copy(kv, ks.buf_.data() + i * 8);
       ks.views_.emplace_back(ks.buf_.data() + i * 8, 8);
